El bono real se calcula sacando el aprovechamiento individual, topandolo al 100% y sacando el promedio del equipo
closes  #4

+++ backend/app/core/models/jugador_model.py
@@ -23,7 +23,7 @@ class JugadorSueldoModel(BaseModel):
     goles: int
     sueldo: int
     bono: int
-    sueldo_completo: int
+    sueldo_completo: float
     equipo: str
